+++ database/migrations/2023_06_14_200520_create_opencajas_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->decimal('totalcash', 10, 2)->default(0);
             $table->decimal('totaltransfer', 10, 2)->default(0);
             $table->enum('status', [Opencaja::ACTIVO, Opencaja::INACTIVO])->default(Opencaja::ACTIVO);
-            $table->bigInteger('caja_id')->nullable();
+            $table->unsignedTinyInteger('caja_id')->nullable();
             $table->bigInteger('user_id')->nullable();
             $table->foreign('caja_id')->on('cajas')->references('id');
             $table->foreign('user_id')->on('users')->references('id');
